requirements.txt
asgiref==3.3.1
gunicorn==19.7.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-cors-headers==3.6.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
idna==2.10
pudb==2020.1
Pygments==2.7.3
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.2
urwid==2.1.2
django-nose==1.4.6
nose-progressive==1.5.2
